requirements.txt
# Core API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Quantized LLM support
ctransformers>=0.2.27

# Optional: For local .env files
python-dotenv>=1.0.0
